requirements.txt
boto3
flask
pyminizip
flask_sqlalchemy
psycopg2-binary
Flask-Migrate
schedule
flask_httpauth
pyminizip